refactor hot handler
